Add a link to add a project in things
This change adds a link at the top of the post to add a project to Things. The link is just a things:// link that will make a project with the checklist on the github repo. It makes it easier to quickly make a whole project with the entire checklist in it. 

More on things url scheme here: https://culturedcode.com/things/support/articles/2803573/